Unsubscribe feature added, new subscription, edit subscription and options screens added

+++ app/src/main/java/ru/ermakov/creator/data/remote/model/RemoteSubscription.kt
@@ -12,5 +12,5 @@ data class RemoteSubscription(
     @SerializedName("description")
     val description: String,
     @SerializedName("price")
-    val price: Int
+    val price: Long
 )
