+++ app/src/main/java/ru/ermakov/creator/data/remote/model/RemoteSubscriptionRequest.kt
@@ -11,5 +11,5 @@ data class RemoteSubscriptionRequest(
     @SerializedName("description")
     val description: String,
     @SerializedName("price")
-    val price: Int
+    val price: Long
 )
