+++ app/src/main/java/ru/ermakov/creator/domain/model/SubscriptionRequest.kt
@@ -4,5 +4,5 @@ data class SubscriptionRequest(
     val creatorId: String,
     val title: String,
     val description: String,
-    val price: Int
+    val price: Long
 )
